requirements.txt
google-generativeai==0.7.0
opencv-python==4.10.0.84
pillow==10.4.0
Twisted==24.3.0